move tests around, rename, add more schedule func tests
